+++ HISTORY.rst
@@ -10,6 +10,7 @@ Release History
 
 - Documented how to use Jupytext as a Python library (#317)
 - Mention of the server extension in the documentation (#304)
+- Test a `README.md` is as simple as `jupytext --execute README.md` (#303)
 
 
 1.2.1 (2019-07-20)